#5566: Move setMapTriggerEpic to epics/identify.js (#5754)
