+++ web/client/epics/__tests__/identify-test.js
@@ -8,7 +8,7 @@
 
 const expect = require('expect');
 const { LOCATION_CHANGE } = require('connected-react-router');
-const { ZOOM_TO_POINT, clickOnMap, CHANGE_MAP_VIEW, UNREGISTER_EVENT_LISTENER} = require('../../actions/map');
+const { ZOOM_TO_POINT, clickOnMap, CHANGE_MAP_VIEW, UNREGISTER_EVENT_LISTENER, REGISTER_EVENT_LISTENER} = require('../../actions/map');
 const { FEATURE_INFO_CLICK, UPDATE_CENTER_TO_MARKER, PURGE_MAPINFO_RESULTS, NEW_MAPINFO_REQUEST, LOAD_FEATURE_INFO, NO_QUERYABLE_LAYERS, ERROR_FEATURE_INFO, EXCEPTIONS_FEATURE_INFO, SHOW_MAPINFO_MARKER, HIDE_MAPINFO_MARKER, GET_VECTOR_INFO, SET_CURRENT_EDIT_FEATURE_QUERY, loadFeatureInfo, featureInfoClick, closeIdentify, toggleHighlightFeature, editLayerFeatures, updateFeatureInfoClickPoint, purgeMapInfoResults } = require('../../actions/mapInfo');
 const { REMOVE_MAP_POPUP } = require('../../actions/mapPopups');
 
@@ -24,13 +24,15 @@ const {
     hideMarkerOnIdentifyClose,
     onUpdateFeatureInfoClickPoint,
     removePopupOnLocationChangeEpic,
-    removePopupOnUnregister
+    removePopupOnUnregister,
+    setMapTriggerEpic
 } = require('../identify').default;
 const { CLOSE_ANNOTATIONS } = require('../../actions/annotations');
 const { testEpic, TEST_TIMEOUT, addTimeoutEpic } = require('./epicTestUtils');
 const { registerHook, GET_COORDINATES_FROM_PIXEL_HOOK, GET_PIXEL_FROM_COORDINATES_HOOK } = require('../../utils/MapUtils');
 const { setControlProperties } = require('../../actions/controls');
 const { BROWSE_DATA } = require('../../actions/layers');
+const { configureMap } = require('../../actions/config');
 
 const TEST_MAP_STATE = {
     present: {
@@ -924,6 +926,22 @@ describe('identify Epics', () => {
                     popups: [{id: "id"}]
                 }
             });
+        });
+    });
+    describe('setMapTriggerEpic', () => {
+        it('should register event if hover is trigger in mapInfo', (done) => {
+            const epicResponse = actions => {
+                expect(actions[0].type).toBe(REGISTER_EVENT_LISTENER);
+                done();
+            };
+            testEpic(setMapTriggerEpic, 1, configureMap(), epicResponse, {mapInfo: {configuration: {trigger: 'hover'}}});
+        });
+        it('should unregister event if no mapInfo or no trigger is present in mapInfo', (done) => {
+            const epicResponse = actions => {
+                expect(actions[0].type).toBe(UNREGISTER_EVENT_LISTENER);
+                done();
+            };
+            testEpic(setMapTriggerEpic, 1, configureMap(), epicResponse, {});
         });
     });
 });
